Dinamical manipulation of filters of comments + no name repetition allowed in models and products adition

+++ controllers/usersController.php
@@ -1,20 +1,34 @@
 <?php
     require_once('./models/usersModel.php');
-    // require_once('./views/usersView.php');
+    require_once('./views/usersView.php');
+    require_once('loginController.php');
     // I mean what is going to be the new usersView
 
     class UsersController{
 
         private $model;
         private $view;
+        private $log;
         public function __construct(){
             $this->model = new UsersModel();
-            // $this->view = new UsersView();
+            $this->view = new UsersView();
+            $this->log = new LoginController();
+            // session_start();
+
         }
 
-        public function getUsers(){
-            $users = $this->model->getUsers();
-            $this->view->showUsers($users);
+        public function checkAdmin(){
+            return($this->log->isAdmin());
+        }
+
+        public function commandUsers(){
+            if($this->checkAdmin()){
+                $users = $this->model->getUsers();
+                $this->view->showUsers($users);
+            }
+            else{
+                header("Location: ".URL);
+            }
         }
 
         public function findUser($userName){
@@ -22,14 +36,34 @@
         }
 
         public function addUser($user){
-            $this->model->$this->model->addUser($user);
+            if($this->checkAdmin()){
+                $this->model->addUser($user);
+            }
+            else{
+                header("Location: ".URL);
+            }
         }
 
-        public function deleteUser($id){
-            $this->model->deleteUser($id);
+        public function deleteUser($params = null){
+            if($this->checkAdmin()){
+                $id = $params[':ID'];
+                $this->model->deleteUser($id);
+                header("Location: ".URL."usuarios");
+            }
+            else{
+                header("Location: ".URL);
+            }
         }
 
-        public function toogleAdmin($id){
-            $this->model->toogleAdmin($id);
+        public function toggleAdmin($params = null){
+            // var_dump($id); die;
+            if($this->checkAdmin()){
+                $id = $params[':ID'];
+                $this->model->toggleAdmin($id);
+                header("Location: ".URL."usuarios");
+            }
+            else{
+                header("Location: ".URL);
+            }
         }
 }
